Use runner id in the filename of the config file
The combo runner_target_type-runner_target does not work when one needs
to define two runners with different parameters for a same group. Use
the runner ID which is unique.

Not sure how the role will behave if a runner is removed from GitLab
while its configuration snippet is still in the local file system.